refactor(uzicli): split schedule verbs into client_schedules.go/fake_schedules.go (PRD #1017 M2)
Pure motion: move the 14 schedule HTTPClient verbs and their FakeClient twins
into per-domain files. The floating // StreamRun replays... doc comment (bound by
adjacency to ListSchedules) travels with ListSchedules to preserve go-doc output.
go doc -u -all byte-identical; no test edits.
